Device identification in threading mode.
